pyproject.toml
[tool.poetry]
name = "financial-analysis"
version = "0.1.0"
description = "Financial Analysis Chat Agent — SEC EDGAR + Claude"
authors = ["nimohunter <nimohunter@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
pgvector = "^0.3.5"
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
anthropic = "^0.40.0"
groq = ">=0.13.0"
# fastembed installed separately: pip install fastembed==0.3.6 (poetry can't resolve on Python 3.9)
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
tiktoken = "^0.8.0"
httpx = "^0.28.1"
tenacity = "^9.0.0"
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.ml.dependencies]
# Kept for reference — replaced by fastembed in main deps (no torch required)

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
